lib-src:
2009-12-09  Alexey Zakhlestin <indeyets@gmail.com>

	* lib-src/pointer-set.c: Add const markers.
	* lib-src/pointer-set.h: Add const markers.

libgst:
2009-12-09  Alexey Zakhlestin <indeyets@gmail.com>

	* libgst/gstpriv.h: Include sockets.h.
	* libgst/print.c: Fix printf.
	* libgst/socklen.c: Use socklen_t where appropriate.
	* libgst/sym.c: Remove useless cast.
	* libgst/sysdep/common/files.c: Fix 64-bit cleanliness problems.
	* libgst/sysdep/posix/findexec.c: Likewise.

packages/blox/tk:
2009-12-09  Alexey Zakhlestin <indeyets@gmail.com>

	* BloxTK.c: Fix NULL vs. 0.

+++ libgst/sysdep/posix/findexec.c
@@ -171,7 +171,7 @@ find_executable (const char *argv0)
 	  else if (!p_next)
 	    asprintf (&concat_name, "%s/%s", p, argv0);
 	  else
-	    asprintf (&concat_name, "%.*s/%s", p_next++ - p, p, argv0);
+	    asprintf (&concat_name, "%.*s/%s", (int)(p_next++ - p), p, argv0);
 
 	  if (maybe_executable (concat_name))
 	    {
